iota_grpc_server::server

Function start_grpc_server

pub async fn start_grpc_server(
    grpc_reader: Arc<GrpcReader>,
    executor: Option<Arc<dyn TransactionExecutor>>,
    config: GrpcApiConfig,
    shutdown_token: CancellationToken,
    chain_id: ChainIdentifier,
    metrics: Option<GrpcServerMetrics>,
) -> Result<GrpcServerHandle>

Start a gRPC server with checkpoint and event services

This function creates and starts a gRPC server that hosts checkpoint-related and event streaming services. Currently includes the checkpoint streaming and event streaming services, but can be extended to host additional services in the future.